sinkv2(ticdc): move mysql_params.go into mysql package
sinkv2(ticdc): add config

sinkv2(ticdc): better code
