+++ cdc/sinkv2/eventsink/txn/mysql/mysql.go
@@ -18,46 +18,36 @@ import (
 	"database/sql"
 	"database/sql/driver"
 	"fmt"
-	"net"
 	"net/url"
 	"sort"
-	"strconv"
 	"time"
 
 	dmysql "github.com/go-sql-driver/mysql"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/log"
-	"github.com/pingcap/tidb/parser/charset"
 	"github.com/pingcap/tidb/parser/mysql"
 	"github.com/pingcap/tiflow/cdc/model"
 	"github.com/pingcap/tiflow/cdc/sinkv2/eventsink"
 	"github.com/pingcap/tiflow/cdc/sinkv2/metrics"
-	dmutils "github.com/pingcap/tiflow/dm/pkg/utils"
+	"github.com/pingcap/tiflow/pkg/config"
 	cerror "github.com/pingcap/tiflow/pkg/errors"
 	"github.com/pingcap/tiflow/pkg/quotes"
 	"github.com/pingcap/tiflow/pkg/retry"
 	"github.com/pingcap/tiflow/pkg/sink"
+	pmysql "github.com/pingcap/tiflow/pkg/sink/mysql"
 	"go.uber.org/zap"
 )
 
 const (
-	backoffBaseDelay = 500 * time.Millisecond
-	backoffMaxDelay  = 60 * time.Second
-
 	// Max interval for flushing transactions to the downstream.
 	maxFlushInterval = 100 * time.Millisecond
-
-	defaultDMLMaxRetry uint64 = 8
-
-	// defaultDDLMaxRetry uint64 = 20
 )
 
 type mysqlBackend struct {
 	changefeedID model.ChangeFeedID
 	db           *sql.DB
-	params       *sinkParams
-	options      SinkOptions
+	cfg          *pmysql.Config
 	statistics   *metrics.Statistics
 
 	events []*eventsink.TxnCallbackableEvent
@@ -71,120 +61,48 @@ func NewMySQLBackend(
 	ctx context.Context,
 	changefeedID model.ChangeFeedID,
 	sinkURI *url.URL,
-	opts SinkOptions,
+	replicaConfig *config.ReplicaConfig,
+	dbConnFactory pmysql.Factory,
 ) (*mysqlBackend, error) {
-	params, err := parseSinkURIToParams(ctx, changefeedID, sinkURI)
+	cfg := pmysql.NewConfig()
+	err := cfg.Apply(ctx, changefeedID, sinkURI, replicaConfig)
 	if err != nil {
 		return nil, err
 	}
 
-	dsnStr, err := adjustDSN(ctx, sinkURI, params, opts)
+	dsnStr, err := pmysql.GenerateDSN(ctx, sinkURI, cfg, dbConnFactory)
 	if err != nil {
 		return nil, err
 	}
 
-	db, err := opts.getDBConn(ctx, dsnStr)
+	db, err := dbConnFactory(ctx, dsnStr)
 	if err != nil {
 		return nil, err
 	}
-	db.SetMaxIdleConns(params.workerCount)
-	db.SetMaxOpenConns(params.workerCount)
+	db.SetMaxIdleConns(cfg.WorkerCount)
+	db.SetMaxOpenConns(cfg.WorkerCount)
 
 	ctx, cancel := context.WithCancel(ctx)
-	sink := &mysqlBackend{
+	s := &mysqlBackend{
 		changefeedID: changefeedID,
 		db:           db,
-		params:       params,
-		options:      opts,
+		cfg:          cfg,
 		statistics:   metrics.NewStatistics(ctx, sink.TxnSink),
 		cancel:       cancel,
 	}
 
 	log.Info("mysql backend is created",
 		zap.String("changefeedID", fmt.Sprintf("%s.%s", changefeedID.Namespace, changefeedID.ID)),
-		zap.Bool("forceReplicate", sink.options.forceReplicate),
-		zap.Bool("enableOldValue", sink.options.enableOldValue))
-	return sink, nil
-}
-
-func adjustDSN(ctx context.Context, sinkURI *url.URL, params *sinkParams, opts SinkOptions) (dsnStr string, err error) {
-	// dsn format of the driver:
-	// [username[:password]@][protocol[(address)]]/dbname[?param1=value1&...&paramN=valueN]
-	username := sinkURI.User.Username()
-	if username == "" {
-		username = "root"
-	}
-	password, _ := sinkURI.User.Password()
-	hostName := sinkURI.Hostname()
-	port := sinkURI.Port()
-	if port == "" {
-		port = "4000"
-	}
-
-	// This will handle the IPv6 address format.
-	var dsn *dmysql.Config
-	host := net.JoinHostPort(hostName, port)
-	dsnStr = fmt.Sprintf("%s:%s@tcp(%s)/%s", username, password, host, params.tls)
-	if dsn, err = dmysql.ParseDSN(dsnStr); err != nil {
-		return
-	}
-
-	// create test db used for parameter detection
-	// Refer https://github.com/go-sql-driver/mysql#parameters
-	if dsn.Params == nil {
-		dsn.Params = make(map[string]string, 1)
-	}
-	if params.timezone != "" {
-		dsn.Params["time_zone"] = params.timezone
-	}
-	dsn.Params["readTimeout"] = params.readTimeout
-	dsn.Params["writeTimeout"] = params.writeTimeout
-	dsn.Params["timeout"] = params.dialTimeout
-
-	var testDB *sql.DB
-	testDB, err = opts.getDBConn(ctx, dsn.FormatDSN())
-	if err != nil {
-		return
-	}
-	defer testDB.Close()
-
-	// Adjust sql_mode for compatibility.
-	dsn.Params["sql_mode"], err = querySQLMode(ctx, testDB)
-	if err != nil {
-		return
-	}
-	dsn.Params["sql_mode"], err = dmutils.AdjustSQLModeCompatible(dsn.Params["sql_mode"])
-	if err != nil {
-		return
-	}
-	// NOTE: quote the string is necessary to avoid ambiguities.
-	dsn.Params["sql_mode"] = strconv.Quote(dsn.Params["sql_mode"])
-
-	dsnStr, err = generateDSNByParams(ctx, dsn, params, testDB)
-	if err != nil {
-		return
-	}
-
-	// check if GBK charset is supported by downstream
-	var gbkSupported bool
-	gbkSupported, err = checkCharsetSupport(ctx, testDB, charset.CharsetGBK)
-	if err != nil {
-		return
-	}
-	if !gbkSupported {
-		log.Warn("GBK charset is not supported by the downstream. "+
-			"Some types of DDLs may fail to execute",
-			zap.String("hostname", hostName), zap.String("port", port))
-	}
-
-	return
+		zap.Bool("forceReplicate", s.cfg.ForceReplicate),
+		zap.Bool("enableOldValue", s.cfg.EnableOldValue))
+	return s, nil
 }
 
 // OnTxnEvent implements interface backend.
 func (s *mysqlBackend) OnTxnEvent(event *eventsink.TxnCallbackableEvent) (needFlush bool) {
 	s.events = append(s.events, event)
 	s.rows += len(event.Event.Rows)
-	return event.Event.ToWaitFlush() || s.rows >= s.params.maxTxnRow
+	return event.Event.ToWaitFlush() || s.rows >= s.cfg.MaxTxnRow
 }
 
 // Flush implements interface backend.
@@ -243,35 +161,12 @@ func (s *mysqlBackend) MaxFlushInterval() time.Duration {
 	return maxFlushInterval
 }
 
-func querySQLMode(ctx context.Context, db *sql.DB) (sqlMode string, err error) {
-	row := db.QueryRowContext(ctx, "SELECT @@SESSION.sql_mode;")
-	err = row.Scan(&sqlMode)
-	if err != nil {
-		err = cerror.WrapError(cerror.ErrMySQLQueryError, err)
-	}
-	return
-}
-
-// check whether the target charset is supported
-func checkCharsetSupport(ctx context.Context, db *sql.DB, charsetName string) (bool, error) {
-	// validate charsetName
-	_, err := charset.GetCharsetInfo(charsetName)
-	if err != nil {
-		return false, errors.Trace(err)
-	}
-
-	var characterSetName string
-	querySQL := "select character_set_name from information_schema.character_sets " +
-		"where character_set_name = '" + charsetName + "';"
-	err = db.QueryRowContext(ctx, querySQL).Scan(&characterSetName)
-	if err != nil && err != sql.ErrNoRows {
-		return false, cerror.WrapError(cerror.ErrMySQLQueryError, err)
-	}
-	if err != nil {
-		return false, nil
-	}
-
-	return true, nil
+type preparedDMLs struct {
+	startTs   []model.Ts
+	sqls      []string
+	values    [][]interface{}
+	callbacks []eventsink.CallbackFunc
+	rowCount  int
 }
 
 // prepareDMLs converts model.RowChangedEvent list to query string list and args list
@@ -285,8 +180,8 @@ func (s *mysqlBackend) prepareDMLs() *preparedDMLs {
 
 	// flush cached batch replace or insert, to keep the sequence of DMLs
 	flushCacheDMLs := func() {
-		if s.params.batchReplaceEnabled && len(replaces) > 0 {
-			replaceSqls, replaceValues := reduceReplace(replaces, s.params.batchReplaceSize)
+		if s.cfg.BatchReplaceEnabled && len(replaces) > 0 {
+			replaceSqls, replaceValues := reduceReplace(replaces, s.cfg.BatchReplaceSize)
 			sqls = append(sqls, replaceSqls...)
 			values = append(values, replaceValues...)
 			replaces = make(map[string][][]interface{})
@@ -294,7 +189,7 @@ func (s *mysqlBackend) prepareDMLs() *preparedDMLs {
 	}
 
 	// translateToInsert control the update and insert behavior
-	translateToInsert := s.options.enableOldValue && !s.params.safeMode
+	translateToInsert := s.cfg.EnableOldValue && !s.cfg.SafeMode
 	for _, event := range s.events {
 		for _, row := range event.Event.Rows {
 			if !translateToInsert {
@@ -324,7 +219,7 @@ func (s *mysqlBackend) prepareDMLs() *preparedDMLs {
 			// NOTICE: Only update events with the old value feature enabled will have both columns and preColumns.
 			if translateToInsert && len(row.PreColumns) != 0 && len(row.Columns) != 0 {
 				flushCacheDMLs()
-				query, args = prepareUpdate(quoteTable, row.PreColumns, row.Columns, s.options.forceReplicate)
+				query, args = prepareUpdate(quoteTable, row.PreColumns, row.Columns, s.cfg.ForceReplicate)
 				if query != "" {
 					sqls = append(sqls, query)
 					values = append(values, args)
@@ -341,7 +236,7 @@ func (s *mysqlBackend) prepareDMLs() *preparedDMLs {
 			// It will be translated directly into a DELETE SQL.
 			if len(row.PreColumns) != 0 {
 				flushCacheDMLs()
-				query, args = prepareDelete(quoteTable, row.PreColumns, s.options.forceReplicate)
+				query, args = prepareDelete(quoteTable, row.PreColumns, s.cfg.ForceReplicate)
 				if query != "" {
 					sqls = append(sqls, query)
 					values = append(values, args)
@@ -358,7 +253,7 @@ func (s *mysqlBackend) prepareDMLs() *preparedDMLs {
 			// INSERT(old value is enabled and not in safe mode)
 			// or REPLACE(old value is disabled or in safe mode) SQL.
 			if len(row.Columns) != 0 {
-				if s.params.batchReplaceEnabled {
+				if s.cfg.BatchReplaceEnabled {
 					query, args = prepareReplace(quoteTable, row.Columns, false /* appendPlaceHolder */, translateToInsert)
 					if query != "" {
 						if _, ok := replaces[query]; !ok {
@@ -397,44 +292,6 @@ func (s *mysqlBackend) prepareDMLs() *preparedDMLs {
 		callbacks: callbacks,
 		rowCount:  rowCount,
 	}
-}
-
-func logDMLTxnErr(
-	err error, start time.Time, changefeed model.ChangeFeedID,
-	query string, count int, startTs []model.Ts,
-) error {
-	if isRetryableDMLError(err) {
-		log.Warn("execute DMLs with error, retry later",
-			zap.Error(err), zap.Duration("duration", time.Since(start)),
-			zap.String("query", query), zap.Int("count", count),
-			zap.Uint64s("startTs", startTs),
-			zap.String("namespace", changefeed.Namespace),
-			zap.String("changefeed", changefeed.ID))
-	} else {
-		log.Error("execute DMLs with error, can not retry",
-			zap.Error(err), zap.Duration("duration", time.Since(start)),
-			zap.String("query", query), zap.Int("count", count),
-			zap.String("namespace", changefeed.Namespace),
-			zap.String("changefeed", changefeed.ID))
-	}
-	return err
-}
-
-func isRetryableDMLError(err error) bool {
-	if !cerror.IsRetryableError(err) {
-		return false
-	}
-
-	errCode, ok := getSQLErrCode(err)
-	if !ok {
-		return true
-	}
-
-	switch errCode {
-	case mysql.ErrNoSuchTable, mysql.ErrBadDB:
-		return false
-	}
-	return true
 }
 
 func (s *mysqlBackend) execDMLWithMaxRetries(ctx context.Context, dmls *preparedDMLs) error {
@@ -497,18 +354,48 @@ func (s *mysqlBackend) execDMLWithMaxRetries(ctx context.Context, dmls *prepared
 			zap.String("changefeed", s.changefeedID.ID),
 			zap.Int("numOfRows", dmls.rowCount))
 		return nil
-	}, retry.WithBackoffBaseDelay(backoffBaseDelay.Milliseconds()),
-		retry.WithBackoffMaxDelay(backoffMaxDelay.Milliseconds()),
-		retry.WithMaxTries(s.options.dmlMaxRetry),
+	}, retry.WithBackoffBaseDelay(pmysql.BackoffBaseDelay.Milliseconds()),
+		retry.WithBackoffMaxDelay(pmysql.BackoffMaxDelay.Milliseconds()),
+		retry.WithMaxTries(s.cfg.DMLMaxRetry),
 		retry.WithIsRetryableErr(isRetryableDMLError))
 }
 
-type preparedDMLs struct {
-	startTs   []model.Ts
-	sqls      []string
-	values    [][]interface{}
-	callbacks []eventsink.CallbackFunc
-	rowCount  int
+func logDMLTxnErr(
+	err error, start time.Time, changefeed model.ChangeFeedID,
+	query string, count int, startTs []model.Ts,
+) error {
+	if isRetryableDMLError(err) {
+		log.Warn("execute DMLs with error, retry later",
+			zap.Error(err), zap.Duration("duration", time.Since(start)),
+			zap.String("query", query), zap.Int("count", count),
+			zap.Uint64s("startTs", startTs),
+			zap.String("namespace", changefeed.Namespace),
+			zap.String("changefeed", changefeed.ID))
+	} else {
+		log.Error("execute DMLs with error, can not retry",
+			zap.Error(err), zap.Duration("duration", time.Since(start)),
+			zap.String("query", query), zap.Int("count", count),
+			zap.String("namespace", changefeed.Namespace),
+			zap.String("changefeed", changefeed.ID))
+	}
+	return err
+}
+
+func isRetryableDMLError(err error) bool {
+	if !cerror.IsRetryableError(err) {
+		return false
+	}
+
+	errCode, ok := getSQLErrCode(err)
+	if !ok {
+		return true
+	}
+
+	switch errCode {
+	case mysql.ErrNoSuchTable, mysql.ErrBadDB:
+		return false
+	}
+	return true
 }
 
 func getSQLErrCode(err error) (errors.ErrCode, bool) {
@@ -518,20 +405,4 @@ func getSQLErrCode(err error) (errors.ErrCode, bool) {
 	}
 
 	return errors.ErrCode(mysqlErr.Number), true
-}
-
-func getDBConn(ctx context.Context, dsnStr string) (*sql.DB, error) {
-	db, err := sql.Open("mysql", dsnStr)
-	if err != nil {
-		return nil, cerror.ErrMySQLConnectionError.Wrap(err).GenWithStack("fail to open MySQL connection")
-	}
-	err = db.PingContext(ctx)
-	if err != nil {
-		// close db to recycle resources
-		if closeErr := db.Close(); closeErr != nil {
-			log.Warn("close db failed", zap.Error(err))
-		}
-		return nil, cerror.ErrMySQLConnectionError.Wrap(err).GenWithStack("fail to open MySQL connection")
-	}
-	return db, nil
 }
